+++ lib/src/transformation/resize/fit.dart
@@ -1,24 +1,7 @@
-// package com.cloudinary.transformation.resize
-//
-// class Fit(
-// dimensions: Dimensions, relative: Boolean? = null,
-// regionRelative: Boolean? = null
-// ) :
-// Resize(dimensions, relative, regionRelative) {
-//   override val actionType = "fit"
-//
-//   class Builder : BaseBuilder<Builder>() {
-//   override fun getThis() = this
-//   override fun build() = Fit(
-//   Dimensions(width, height, aspectRatio),
-//   relative,
-//   regionRelative
-//   )
-//   }
-// }
-
 import 'package:cloudinary_dart/src/transformation/common.dart';
 import 'package:cloudinary_dart/src/transformation/resize/resize.dart';
+
+import 'common.dart';
 
 /// Class Fit
 class Fit extends Resize {
